+++ packages/admin/src/admin.scss
@@ -3,6 +3,7 @@
 body {
   font-family: "Roboto", sans-serif;
   line-height: initial;
+  height: auto;
 }
 
 .ql-editor {
