+++ packages/admin/src/pages/BoardPage/ArticleList/ArticleList.view.tsx
@@ -18,10 +18,10 @@ export default function ArticleListView({ articles, onClickArticle }: Props) {
       <div className={classnames($.row, $.header)}>
         <div className={classnames($.cell, $.id)}>ID</div>
         <div className={classnames($.cell, $.title)}>제목</div>
-        <div className={classnames($.cell, $.date)}>작성일</div>
         <div className={classnames($.cell, $.id)}>조회수</div>
         <div className={classnames($.cell, $.scraps)}>스크랩</div>
-        <div className={classnames($.cell, $.id)}>수정일</div>
+        <div className={classnames($.cell, $.date)}>작성</div>
+        <div className={classnames($.cell, $.date)}>수정</div>
       </div>
       {articles.map((article) => (
         <div
@@ -32,12 +32,12 @@ export default function ArticleListView({ articles, onClickArticle }: Props) {
           <div className={classnames($.cell, $.id)}>{article.id}</div>
           <div className={classnames($.cell, $.title)}>{article.title}</div>
 
-          <div className={classnames($.cell, $.date)}>
-            {dayjs(article.dateTime).format("YYYY.MM.DD")}
-          </div>
           <div className={classnames($.cell, $.hits)}>{article.viewCount}</div>
           <div className={classnames($.cell, $.scraps)}>
             {article.bookmarkCount}
+          </div>
+          <div className={classnames($.cell, $.date)}>
+            {dayjs(article.dateTime).format("YYYY.MM.DD")}
           </div>
           <div className={classnames($.cell, $.date)}>
             {dayjs(article.updatedDateTime).format("YYYY.MM.DD HH:MM")}
