+++ packages/admin/src/pages/PlaceManagementPage/Place.route.tsx
@@ -2,16 +2,16 @@ import { Route, Routes } from "react-router-dom";
 
 import PageLayout from "src/components/Layout/Page/Page.view";
 
-import PlaceAdd from "./PlaceAdd/PlaceAdd";
+import Place from "./Place/Place";
 import PlaceList from "./PlaceList/PlaceList";
 
 export default function PlaceManagementPage() {
   return (
     <PageLayout>
       <Routes>
-        <Route path="/add" element={<PlaceAdd />} />
-        <Route path="/add/:placeId" element={<PlaceAdd />} />
         <Route path="/list" element={<PlaceList />} />
+        <Route path="/add" element={<Place />} />
+        <Route path="/:placeId" element={<Place />} />
       </Routes>
     </PageLayout>
   );
