+++ packages/admin/src/pages/PlaceManagementPage/Place/Place.module.scss
@@ -1,74 +1,16 @@
-@import "@shared/styles/main.scss";
-@import "src/styles/mixin.scss";
+.wrapper {
+    margin-left: 30px;
+    margin-top: 30px;
+    width: 800px;
 
-.article {
-  margin: 20px;
-  padding: 20px;
-  background-color: white;
-  .title {
-    font-size: 25px;
-  }
-
-  .content {
-    margin: 20px 0px;
-  }
-
-  .hr {
-    margin: 10px 0px;
-    background-color: rgb(209, 209, 209);
-    height: 1px;
-    border: 0;
-  }
-
-  .image-list {
-    overflow: auto;
-    display: flex;
-
-    .image-card {
-      width: 150px;
-      height: 150px;
-      margin-right: 15px;
-      cursor: pointer;
-      position: relative;
-      background-color: rgba(0, 0, 0, 0.138);
-
-      .image {
-        background-size: contain;
-        background-repeat: no-repeat;
-        background-position: center;
-        width: 150px;
-        height: 150px;
-        resize: both;
-      }
+    .title {
+        margin-left: 30px;
+        margin-bottom: 20px;
+        font-size: x-large;
     }
-  }
-}
 
-.button {
-  @include no-line;
-  width: 5.5rem;
-  min-height: 2.3rem;
-  border-radius: 0.3rem;
-  background-color: #005299;
-  color: $white;
-  font-size: 1.1rem;
-  font-weight: 500;
-  margin-top: 20px;
-
-  &:hover {
-    background-color: $blue-200;
-  }
-
-  &:first-child {
-    margin-right: 10px;
-  }
-}
-
-.text {
-    font-size: 16px;
-}
-
-.bold {
-    font-size: 16px;
-    font-weight: 600;
+    .submit {
+        display: flex;
+        justify-content: center;
+    }
 }
